apps/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tripflow-api"
version = "0.1.0"
description = "TripFlow FastAPI backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "supabase>=2.10.0",
    "PyJWT[crypto]>=2.10.1",
    "httpx>=0.28.0",
    "certifi>=2024.2.0",
    "resend>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "respx>=0.21.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing"

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = []

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
